refactor: 메인 패키지 구조 변경에 따른 테스트 코드 패키지 위치 변경
- MoneyTest 위치 변경
- LottoTicketsTest 테스트 오타 수정
