+++ src/test/java/domain/lotto/LottoTicketsTest.java
@@ -1,6 +1,5 @@
 package domain.lotto;
 
-import domain.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +9,7 @@ public class LottoTicketsTest {
 
     @Test
     @DisplayName("6000원이 주어졌을 때 랜덤한 로또 티켓 6개를 생성한다.")
-    void createRandomLottoTicketsq() {
+    void createRandomLottoTicket() {
         LottoTickets lottoTickets = LottoTickets.createRandomTickets(new Money(6000));
 
         assertThat(lottoTickets.getLottoTickets().size()).isEqualTo(6);
